+++ openmdao/core/explicitcomponent.py
@@ -10,7 +10,7 @@ from openmdao.utils.class_util import overrides_method
 from openmdao.utils.general_utils import ContainsAll
 from openmdao.recorders.recording_iteration_stack import Recording
 from openmdao.core.constants import INT_DTYPE
-from openmdao.warnings import warn_deprecation
+from openmdao.utils.om_warnings import warn_deprecation
 
 _inst_functs = ['compute_jacvec_product']
