+++ openmdao/recorders/recording_manager.py
@@ -4,7 +4,7 @@ RecordingManager class definition.
 import time
 
 from openmdao.utils.general_utils import simple_warning
-from openmdao.warnings import warn_deprecation
+from openmdao.utils.om_warnings import warn_deprecation
 
 try:
     from openmdao.utils.mpi import MPI
