+++ openmdao/recorders/tests/test_sqlite_recorder.py
@@ -28,7 +28,7 @@ from openmdao.utils.assert_utils import assert_near_equal, assert_equal_arrays, 
     assert_warning, assert_no_warning
 from openmdao.utils.general_utils import determine_adder_scaler
 from openmdao.utils.testing_utils import use_tempdirs
-from openmdao.warnings import OMDeprecationWarning
+from openmdao.utils.om_warnings import OMDeprecationWarning
 
 # check that pyoptsparse is installed. if it is, try to use SLSQP.
 from openmdao.utils.general_utils import set_pyoptsparse_opt
