+++ openmdao/utils/assert_utils.py
@@ -14,7 +14,7 @@ from openmdao.core.component import Component
 from openmdao.core.group import Group
 from openmdao.jacobians.dictionary_jacobian import DictionaryJacobian
 from openmdao.utils.general_utils import pad_name
-from openmdao.warnings import warn_deprecation, reset_warning_registry
+from openmdao.utils.om_warnings import warn_deprecation, reset_warning_registry
 
 
 @contextmanager
