+++ openmdao/utils/om.py
@@ -49,7 +49,7 @@ from openmdao.utils.entry_points import _list_installed_setup_parser, _list_inst
         _find_plugins_setup_parser, _find_plugins_exec
 from openmdao.core.component import Component
 from openmdao.utils.general_utils import ignore_errors
-from openmdao.warnings import warn_deprecation
+from openmdao.utils.om_warnings import warn_deprecation
 
 
 def _n2_setup_parser(parser):
