+++ openmdao/utils/units.py
@@ -16,7 +16,7 @@ import os.path
 from collections import OrderedDict
 
 from configparser import RawConfigParser as ConfigParser
-from openmdao.warnings import warn_deprecation
+from openmdao.utils.om_warnings import warn_deprecation
 
 # pylint: disable=E0611, F0401
 from math import floor, pi
